requirements.txt
pywin32
tkinterdnd2
pillow